production/demo_ex/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile '.\requirements.in'
#
neat-python==0.92
    # via -r .\requirements.in
numpy==1.24.0
    # via -r .\requirements.in
pygame==2.1.2
    # via -r .\requirements.in
pandas~=1.5.2